backend/requirements.txt
Package                  Version
------------------------ -----------
aiohappyeyeballs         2.6.1
aiohttp                  3.13.3
aiosignal                1.4.0
annotated-doc            0.0.4
annotated-types          0.7.0
anyio                    4.12.1
appnope                  0.1.4
asttokens                3.0.1
attrs                    25.4.0
beautifulsoup4           4.14.3
certifi                  2026.1.4
charset-normalizer       3.4.4
cianparser               1.0.4
click                    8.3.1
cloudscraper             1.2.71
comm                     0.2.3
contourpy                1.3.3
cycler                   0.12.1
dataclasses-json         0.6.7
datetime                 6.0
debugpy                  1.8.19
decorator                5.2.1
distro                   1.9.0
dotenv                   0.9.9
executing                2.2.1
fastapi                  0.128.0
filelock                 3.20.3
fonttools                4.61.1
frozenlist               1.8.0
fsspec                   2026.1.0
groq                     0.37.1
grpcio                   1.76.0
h11                      0.16.0
h2                       4.3.0
hf-xet                   1.2.0
hpack                    4.1.0
httpcore                 1.0.9
httpx                    0.28.1
httpx-sse                0.4.3
huggingface-hub          0.36.0
hyperframe               6.1.0
idna                     3.11
ipykernel                7.1.0
ipython                  9.9.0
ipython-pygments-lexers  1.1.1
jedi                     0.19.2
jinja2                   3.1.6
joblib                   1.5.3
jsonpatch                1.33
jsonpointer              3.0.0
jupyter-client           8.8.0
jupyter-core             5.9.1
kiwisolver               1.4.9
langchain                1.2.7
langchain-classic        1.0.1
langchain-community      0.4.1
langchain-core           1.2.7
langchain-groq           1.1.1
langchain-huggingface    1.2.0
langchain-qdrant         1.1.0
langchain-text-splitters 1.1.0
langgraph                1.0.7
langgraph-checkpoint     4.0.0
langgraph-prebuilt       1.0.7
langgraph-sdk            0.3.3
langsmith                0.6.4
lxml                     6.0.2
markupsafe               3.0.3
marshmallow              3.26.2
matplotlib               3.10.8
matplotlib-inline        0.2.1
mpmath                   1.3.0
multidict                6.7.1
mypy-extensions          1.1.0
nest-asyncio             1.6.0
networkx                 3.6.1
numpy                    2.4.1
opencv-python            4.13.0.90
orjson                   3.11.5
ormsgpack                1.12.2
packaging                25.0
pandas                   3.0.0
parso                    0.8.5
pexpect                  4.9.0
pillow                   12.1.0
pip                      25.3
platformdirs             4.5.1
polars                   1.37.1
polars-runtime-32        1.37.1
portalocker              3.2.0
prompt-toolkit           3.0.52
propcache                0.4.1
protobuf                 6.33.4
psutil                   7.2.1
ptyprocess               0.7.0
pure-eval                0.2.3
pydantic                 2.12.5
pydantic-core            2.41.5
pydantic-settings        2.12.0
pygments                 2.19.2
pyparsing                3.3.2
python-dateutil          2.9.0.post0
python-dotenv            1.2.1
python-multipart         0.0.22
pytz                     2025.2
pyyaml                   6.0.3
pyzmq                    27.1.0
qdrant-client            1.16.2
regex                    2026.1.15
requests                 2.32.5
requests-toolbelt        1.0.0
safetensors              0.7.0
scikit-learn             1.8.0
scipy                    1.17.0
sentence-transformers    5.2.0
setuptools               80.10.2
shellingham              1.5.4
six                      1.17.0
sniffio                  1.3.1
soupsieve                2.8.3
sqlalchemy               2.0.46
stack-data               0.6.3
starlette                0.50.0
sympy                    1.14.0
tenacity                 9.1.2
threadpoolctl            3.6.0
tokenizers               0.22.2
torch                    2.10.0
torchaudio               2.10.0
torchvision              0.25.0
tornado                  6.5.4
tqdm                     4.67.1
traitlets                5.14.3
transformers             4.57.6
transliterate            1.10.2
typer-slim               0.21.1
typing-extensions        4.15.0
typing-inspect           0.9.0
typing-inspection        0.4.2
ultralytics              8.4.7
ultralytics-thop         2.0.18
urllib3                  2.6.3
uuid-utils               0.14.0
uvicorn                  0.40.0
wcwidth                  0.4.0
xxhash                   3.6.0
yarl                     1.22.0
zope-interface           8.2
zstandard                0.25.0
